# packages/contracts/openapi/v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/organizations/{organizationAlias}/tournaments/{tournamentAlias}/registrations": {
      "get": {
        "description": "Scoped to the named tournament; another tournament’s entrants are never included.",
        "operationId": "RegistrationsController_list",
        "parameters": [
          {
            "name": "organizationAlias",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tournamentAlias",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistrationResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List a tournament’s registrations",
        "tags": [
          "registrations"
        ]
      }
    },
    "/organizations/{organizationAlias}/tournaments/{tournamentAlias}/registrations/{entrantId}/review": {
      "post": {
        "description": "Audited individually with prior and resulting state.",
        "operationId": "RegistrationsController_review",
        "parameters": [
          {
            "name": "organizationAlias",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tournamentAlias",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entrantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept, refuse or withdraw one registration",
        "tags": [
          "registrations"
        ]
      }
    },
    "/organizations/{organizationAlias}/tournaments/{tournamentAlias}/registrations/bulk-review": {
      "post": {
        "description": "One audit row per registration, never one for the batch: \"approved 40\" does not answer \"why is this team in the draw\".",
        "operationId": "RegistrationsController_bulkReview",
        "parameters": [
          {
            "name": "organizationAlias",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tournamentAlias",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkReviewResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Review several registrations",
        "tags": [
          "registrations"
        ]
      }
    },
    "/organizations/{organizationAlias}/tournaments/{tournamentAlias}/registrations/{entrantId}/team-memberships": {
      "post": {
        "description": "Refused once check-in has closed for a checked-in entrant, whatever a stale console still offers.",
        "operationId": "RegistrationsController_editTeamMemberships",
        "parameters": [
          {
            "name": "organizationAlias",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tournamentAlias",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entrantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditTeamMembershipsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit a registration’s team memberships",
        "tags": [
          "registrations"
        ]
      }
    },
...
  },
  "info": {
    "title": "CopaLibre API",
    "description": "Self-hosted tournament management API.\n\nAuthentication is JWT Bearer, validated against the operator-selected\nidentity provider’s JWKS. Browsers use Authorization Code + PKCE and\nhold the access token in memory only — never localStorage. A token\npassed as a query parameter is ignored, because URLs leak into proxy\nlogs, browser history, metrics, traces and screenshots.\n\nEvery route declares one security plane: public-read,\nauthenticated-interaction, admin-control, or integration.\nauthenticated-interaction and admin-control share this same bearer\ntransport and differ only in authorization policy.",
    "version": "4.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Short-lived access token. Strict stateless mode is the default: no refresh credential is persisted, so a reload reauthenticates."
      }
    },
    "schemas": {
...
      "ProblemResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 403
          },
          "message": {
            "type": "string",
            "example": "subject may only act on their own records"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "TeamMemberResponse": {
        "type": "object",
        "properties": {
          "personId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "example": "Elías Salomón"
          },
          "role": {
            "type": "string",
            "enum": [
              "player",
              "substitute",
              "coach",
              "staff"
            ]
          }
        },
        "required": [
          "personId",
          "displayName",
          "role"
        ]
      },
      "RegistrationResponse": {
        "type": "object",
        "properties": {
          "entrantId": {
            "type": "string",
            "format": "uuid"
          },
          "tournamentId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "refused",
              "withdrawn",
              "checked-in"
            ]
          },
          "teamId": {
            "type": "string",
            "format": "uuid"
          },
          "personId": {
            "type": "string",
            "format": "uuid"
          },
          "teamMembers": {
            "description": "The team entrant’s resulting membership. Populated only by a team-membership edit response.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamMemberResponse"
            }
          }
        },
        "required": [
          "entrantId",
          "tournamentId",
          "status"
        ]
      },
      "ReviewRegistrationRequest": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "accepted",
              "refused",
              "withdrawn"
            ]
          },
          "reason": {
            "type": "string",
            "description": "Recorded on the audit row. A refusal an entrant cannot be told about is one they will ask about."
          }
        },
        "required": [
          "decision"
        ]
      },
      "BulkReviewRequest": {
        "type": "object",
        "properties": {
          "entrantIds": {
            "type": "array",
            "items": {
              "type": "array",
              "format": "uuid"
            }
          },
          "decision": {
            "type": "string",
            "enum": [
              "accepted",
              "refused",
              "withdrawn"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "entrantIds",
          "decision"
        ]
      },
      "BulkReviewResponse": {
        "type": "object",
        "properties": {
          "applied": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistrationResponse"
            }
          },
          "refused": {
            "description": "Registrations left untouched, each with the reason — never silently skipped.",
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        },
        "required": [
          "applied",
          "refused"
        ]
      },
      "EditTeamMembershipsRequest": {
        "type": "object",
        "properties": {
          "personIds": {
            "description": "The team’s full desired membership. Anyone currently a member but not named here is removed.",
            "type": "array",
            "items": {
              "type": "array",
              "format": "uuid"
            }
          }
        },
        "required": [
          "personIds"
        ]
      },
...
    }
  }
}